olga/requirements.txt
django_modeltranslation==0.12.2
Fabric3==1.13.1.post1
Django==1.11.12
future==0.16.0
Mezzanine==4.2.3
Django-oscar==1.5.2
psycopg2==2.7.3.2
gunicorn